rEGISTERATION FORM DONE WITH SOME MORE FUNCTIONALITY IN LOGIN

+++ base/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import redirect, render
 from base.form import Roomform
 from django.db.models import Q
+from django.http import HttpResponse
 from base.models import Room, Topic
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import login , authenticate , logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
 
@@ -15,9 +18,13 @@ from django.contrib.auth import login , authenticate , logout
 # ]
 
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         
         try:
@@ -27,18 +34,35 @@ def loginPage(request):
             
         user = authenticate(request , username=username,password=password)
         
+        
         if user is not None:
             login(request,user)
+            messages.success(request, "You have successfully logged in.")  # Add a success message
             return redirect('home')
         else:
             messages.error(request, "Username or password doesnt exist.")    
-        
-    context = {}
+            
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('login')
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            form.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "There is something Error occured during registation!!!.")    
+    return render(request, 'base/login_register.html', {'form':form})
 
 def homepage(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -57,6 +81,7 @@ def room(request,pk):
     context = {'rooms':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = Roomform()
     if request.method == 'POST':
@@ -64,27 +89,34 @@ def createRoom(request):
         if form.is_valid():
             form.save()
             return redirect('home')
-        
+    
     context = {'form': form}
     return render(request, 'base/room_form.html',context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = Roomform(instance=room)
-    
+        
     if request.method == 'POST':
         form = Roomform(request.POST, instance=room)
         if form.is_valid():
            form.save()
            return redirect('home')
-        
+       
+    if request.user != room.host:
+            return HttpResponse("You are not allowed here!!")
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url="login")
 def deleteForm(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+            return HttpResponse("You are not allowed here!!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
